Convert hooks to selectors.
